+++ src/main.ts
@@ -1,27 +1,25 @@
 import * as fs from "fs/promises";
 import * as core from "@actions/core";
 import * as process from "process";
-import { b2a, exec, moveTo } from "./utils";
-import { Manifest, Status, isNode } from "./types";
+import { exec, moveTo } from "./utils";
+import { Manifest } from "./manifest";
 
 export async function main() {
-  const dbtVersion = core.getInput("dbt-version");
-
   const moveBack = moveTo(core.getInput("dbt-project"));
-  await writeManifest(dbtVersion);
-  const mainManifest = await readManifest("./target/manifest.json");
+  await preprocess();
+  const mainManifest = await Manifest.from("./target/manifest.json");
   moveBack();
 
   let anotherManifest;
   const anotherProject = core.getInput("dbt-project-to-compare-with");
   if (anotherProject) {
     const moveBack = moveTo(anotherProject);
-    await writeManifest(dbtVersion);
-    anotherManifest = await readManifest("./target/manifest.json");
+    await preprocess();
+    anotherManifest = await Manifest.from("./target/manifest.json");
     moveBack();
   }
 
-  const chart = flowchart(mainManifest, anotherManifest);
+  const chart = mainManifest.flowchart(anotherManifest);
   const outpath = `${process.cwd()}/lineage.mermaid`;
   await fs.writeFile(outpath, chart);
   core.setOutput("filepath", outpath);
@@ -44,7 +42,7 @@ const dummyProfile = {
   },
 };
 
-async function writeManifest(dbtVer: string) {
+async function preprocess() {
   const dbtVersion = core.getInput("dbt-version");
   const profiles = "profiles.yml";
 
@@ -63,148 +61,4 @@ async function writeManifest(dbtVer: string) {
   await exec(`pipx run --spec dbt-postgres==${dbtVersion} dbt ls`);
 
   await cleanup();
-}
-
-function readManifest(filepath: string): Promise<Manifest> {
-  return fs
-    .readFile(filepath)
-    .then((buffer) => String(buffer))
-    .then((json) => JSON.parse(json));
-}
-
-export function flowchart(
-  mainManifest: Manifest,
-  anotherManifest?: Manifest,
-): string {
-  const statements = [
-    ...vertices(mainManifest, anotherManifest),
-    ...edges(mainManifest, anotherManifest),
-  ];
-
-  const mermaid =
-    "flowchart LR\n" + statements.map((stmt) => "  " + stmt + ";\n").join("");
-  return mermaid;
-}
-
-function vertices(
-  mainManifest: Manifest,
-  anotherManifest?: Manifest,
-): string[] {
-  let resources: { [key: string]: Status } = {};
-  for (const key of Object.keys({
-    ...mainManifest.sources,
-    ...mainManifest.nodes,
-    ...mainManifest.exposures,
-  })) {
-    resources[key] = "identical";
-  }
-  if (anotherManifest) {
-    for (const key of Object.keys(resources)) {
-      resources[key] = "new";
-    }
-    for (const [key, value] of Object.entries({
-      ...anotherManifest.sources,
-      ...anotherManifest.nodes,
-      ...anotherManifest.exposures,
-    })) {
-      if (key in resources) {
-        if (isNode(value)) {
-          const mainHash = mainManifest.nodes[key].checksum.checksum;
-          const anotherHash = value.checksum.checksum;
-          resources[key] = mainHash === anotherHash ? "identical" : "modified";
-        } else {
-          resources[key] = "identical";
-        }
-      } else {
-        resources[key] = "deleted";
-      }
-    }
-  }
-
-  const statements: string[] = [];
-  for (const [key, value] of Object.entries(resources)) {
-    const splited = key.split(".");
-    let text = splited.slice(2).join(".");
-    const style: string[] = ["color:white", "stroke:black"];
-    switch (splited[0]) {
-      case "source":
-        style.push("fill:green");
-        break;
-      case "seed":
-        style.push("fill:blue");
-        break;
-      case "model":
-        style.push("fill:blue");
-        break;
-      case "exposure":
-        style.push("fill:orange");
-        break;
-    }
-    switch (value) {
-      case "deleted":
-        style.push("stroke-width:4px");
-        style.push("stroke-dasharray: 5 5");
-        break;
-      case "identical":
-        style.push("stroke-width:0px");
-        break;
-      case "modified":
-        style.push("stroke-width:4px");
-        break;
-      case "new":
-        style.push("stroke-width:4px");
-        break;
-    }
-
-    // NOTE
-    // name may contain special character (e.g. white space)
-    // which is not allowed in flowchart id
-    const id = b2a(key);
-    statements.push(`${id}("${text}")`);
-    statements.push(`style ${id} ${style.join(",")}`);
-  }
-  return statements;
-}
-
-function edges(mainManifest: Manifest, anotherManifest?: Manifest): string[] {
-  let mappings: { [key: string]: Status } = {};
-  for (const [parent, children] of Object.entries(mainManifest.child_map)) {
-    for (const child of children) {
-      // since base64 does not use `|`
-      // it is a good separator
-      mappings[`${b2a(parent)}|${b2a(child)}`] = "identical";
-    }
-  }
-  if (anotherManifest) {
-    for (const key of Object.keys(mappings)) {
-      mappings[key] = "new";
-    }
-    for (const [parent, children] of Object.entries(
-      anotherManifest.child_map,
-    )) {
-      for (const child of children) {
-        const key = `${b2a(parent)}|${b2a(child)}`;
-        mappings[key] = key in mappings ? "identical" : "deleted";
-      }
-    }
-  }
-  const statements: string[] = [];
-  let idx = 0;
-  for (const [key, value] of Object.entries(mappings)) {
-    const [parent, child, ..._] = key.split("|");
-    switch (value) {
-      case "deleted":
-        statements.push(`${parent} -.-> ${child}`);
-        break;
-      case "identical":
-        statements.push(`${parent} --> ${child}`);
-        break;
-      case "new":
-        statements.push(`${parent} --> ${child}`);
-        statements.push(`linkStyle ${idx} stroke-width:4px`);
-        break;
-    }
-    idx++;
-  }
-  return statements;
 }
