+++ src/types.ts
@@ -1,20 +1,11 @@
 export type Source = {};
-
-// seed or model
 export type Node = {
   checksum: { checksum: string };
-};
-
+}; // seed or model
 export type Exposure = {};
 
-export type Manifest = {
-  child_map: { [key: string]: string[] };
-  sources: { [key: string]: Source };
-  nodes: { [key: string]: Node };
-  exposures: { [key: string]: Exposure };
-};
-
 export type Resource = Source | Node | Exposure;
+
 export type Status = "identical" | "modified" | "new" | "deleted";
 
 export function isNode(resource: Resource): resource is Node {
